[polybool] -Fix: integer overflow on area calculation with big coords - can cause area sign flip
git-svn-id: svn://repo.hu/librnd/trunk@36683 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/librnd/polybool/pa_pline.c
@@ -235,7 +235,7 @@ void pa_pline_update(rnd_pline_t *pl, rnd_bool optimize)
 	if (c != p) {
 		do {
 			/* calculate area for orientation */
-			area += (double)(p->point[0] - c->point[0]) * (double)(p->point[1] + c->point[1]);
+			area += ((double)p->point[0] - (double)c->point[0]) * ((double)p->point[1] + (double)c->point[1]);
 			pa_pline_box_bump(pl, c->point);
 			pl->Count++;
 		}
